finished authorization

+++ backend/router.js
@@ -1,7 +1,7 @@
 import express from "express";
 import Trip from './Models/trips.js';
 import Experience from './Models/experiences.js';
-// import { checkJwt, authErrorHandler } from './auth.js';
+import { checkJwt } from './auth.js';
 
 const routerTrips = express.Router();
 const routerExperiences = express.Router();
@@ -42,19 +42,6 @@ routerTrips.get('/:id', async (req, res) => {
     }
 });
 
-// View Day of a Trip - probably not needed, delete later
-routerTrips.get('/:tripId/:dayId', async (req, res) => {
-    try {
-        const tripId = req.params.tripId;
-        const dayId = req.params.dayId;
-        const trip = await Trip.findById(tripId);
-        const day = trip.TripDays.find(day => day._id.toString() === dayId);
-       res.status(200).json(day);
-    } catch (error) {
-        console.log(error);
-        res.status(500).json({ 'Error': 'Unable to find day of this trip'});
-    }
-});
 
 // Get all experiences, performs search if there is a query
 routerExperiences.get('/', async (req, res) => {
@@ -97,24 +84,37 @@ routerExperiences.get('/:id', async (req, res) => {
 });
 
 // Private Routes (requires authentication)
-  
+
+// Get Trips created by this user
+routerTrips.get('/user/:userId', checkJwt, async (req, res) => {
+    try {
+        // Extract user ID 
+        const userId = req.auth.sub.split('|')[1];
+
+        // Find trips created by the specified user
+        const trips = await Trip.find({ userId });
+        if (trips.length === 0) {
+            return res.status(404).json({ 'Error': 'No trips found for the specified user' });
+        }
+        res.status(200).json(trips);
+    } catch (error) {
+        console.log(error);
+        res.status(500).json({ 'Error': 'Internal server error' });
+    }
+});
+
 // Create a new Trip
-routerTrips.post('/', async (req, res) => {
+routerTrips.post('/', checkJwt, async (req, res) => {
     try {
         const { TripName, TripDescription, TripImage } = req.body;
-
-        /*
-        uncomment when implementing auth
-        const sub = req.auth.sub;
-        const userId = sub.split('|')[1];
-        console.log(sub);
-        */
+        const userId = req.auth.sub.split('|')[1];
 
         const newTrip = new Trip({
             TripName,
             TripDescription,
             TripDays: [],
-            TripImage
+            TripImage,
+            userId
         });
         
         await newTrip.save();
@@ -127,24 +127,20 @@ routerTrips.post('/', async (req, res) => {
 });
 
 // Add Day to Trip
-routerTrips.post('/:id/add-day', async (req, res) => {
+routerTrips.post('/:id/add-day', checkJwt, async (req, res) => {
     try {
         const tripId = req.params.id;
         const trip = await Trip.findById(tripId);
+        const userId = req.auth.sub.split('|')[1];
 
         if (!trip) {
             return res.status(404).json({ 'Error': 'Trip not found' });
         }
-        
-        /*
-        for auth
-        if (trip.userId !== req.auth.sub.split('|')[1]) {
+        if (trip.userId !== userId) {
             return res.status(403).json({ 'Error': 'You are not authorized to add a day to this trip' });
         }
-        */
 
         const newDayNumber = trip.TripDays.length + 1;
-
         const newDay = {
             DayNumber: newDayNumber,
             DayExperiences: []
@@ -163,20 +159,24 @@ routerTrips.post('/:id/add-day', async (req, res) => {
 });
 
 // Update Trip
-routerTrips.put('/:id', async (req, res) => {
+routerTrips.put('/:id', checkJwt, async (req, res) => {
     try {
+        const userId = req.auth.sub.split('|')[1];
         const { TripName, TripDescription, TripImage } = req.body;
 
-        const updatedTrip = await Trip.findByIdAndUpdate(
-            req.params.id,
-            { TripName, TripDescription, TripImage },
-            { new: true });
-
-        if (updatedTrip) {
-            res.status(200).json(updatedTrip);
-        } else {
-            res.status(404).json({ error: 'Trip not found' });
+        const trip = await Trip.findById(req.params.id);
+        if (!trip) {
+            return res.status(404).json({ 'Error': 'Trip not found' });
         }
+        if (trip.userId !== userId) {
+            return res.status(403).json({ 'Error': 'Forbidden: You are not allowed to edit this trip' });
+        }
+
+        trip.TripName = TripName;
+        trip.TripDescription = TripDescription;
+        trip.TripImage = TripImage;
+        const updatedTrip = await trip.save();
+        res.status(200).json(updatedTrip);
     } catch (error) {
         console.error(error);
         res.status(500).json({ error: 'Unable to update trip' });
@@ -184,8 +184,18 @@ routerTrips.put('/:id', async (req, res) => {
 });
 
 // Delete a Trip
-routerTrips.delete('/:id', async (req, res) => {
+routerTrips.delete('/:id', checkJwt, async (req, res) => {
     try {
+        const trip = await Trip.findById(req.params.id);
+        const userId = req.auth.sub.split('|')[1];
+
+        if (!trip) {
+            return res.status(404).json({ 'Error': 'Trip not found' });
+        }
+        if (trip.userId !== userId) {
+            return res.status(403).json({ 'Error': 'Forbidden: You are not allowed to delete this trip' });
+        }
+
         await Trip.findByIdAndDelete(req.params.id);
         res.status(204).end();
     } catch (error) {
@@ -195,14 +205,18 @@ routerTrips.delete('/:id', async (req, res) => {
 });
 
 // Delete the most recently added Day from a Trip
-routerTrips.delete('/:tripId/delete-day', async (req, res) => {
+routerTrips.delete('/:tripId/delete-day', checkJwt, async (req, res) => {
     try {
         const { tripId } = req.params;
+        const userId = req.auth.sub.split('|')[1];
+
         const trip = await Trip.findById(tripId);
         if (!trip) {
             return res.status(404).json({ 'Error': 'Trip not found' });
         }
-
+        if (trip.userId !== userId) {
+            return res.status(403).json({ 'Error': 'You are not authorized to delete a day from this trip' });
+        }
         if (trip.TripDays.length === 0) {
             return res.status(404).json({ 'Error': 'No days found in the trip' });
         }
@@ -217,14 +231,16 @@ routerTrips.delete('/:tripId/delete-day', async (req, res) => {
 });
 
 // Create a new Experience
-routerExperiences.post('/', async (req, res) => {
+routerExperiences.post('/', checkJwt, async (req, res) => {
     try {
         const { ExperienceName, ExperienceDescription, ExperienceImage } = req.body;
+        const userId = req.auth.sub.split('|')[1];
 
         const newExperience = new Experience({
             ExperienceName,
             ExperienceDescription,
-            ExperienceImage
+            ExperienceImage,
+            userId
         });
 
         await newExperience.save();
@@ -237,29 +253,43 @@ routerExperiences.post('/', async (req, res) => {
 });
 
 // Update Experience
-routerExperiences.put('/:id', async (req, res) => {
+routerExperiences.put('/:id', checkJwt, async (req, res) => {
     try {
+        const userId = req.auth.sub.split('|')[1];
         const { ExperienceName, ExperienceDescription, ExperienceImage } = req.body;
-
-        const updatedExp = await Experience.findByIdAndUpdate(
-            req.params.id,
-            { ExperienceName, ExperienceDescription, ExperienceImage },
-            { new: true });
-
-        if (updatedExp) {
-            res.status(200).json(updatedExp);
-        } else {
-            res.status(404).json({ error: 'Experience not found' });
+        
+        const experience = await Experience.findById(req.params.id);
+        if (!experience) {
+            return res.status(404).json({ 'Error': 'Experience not found' });
         }
+        if (experience.userId !== userId) {
+            return res.status(403).json({ 'Error': 'You are not authorized to update this experience' });
+        }
+
+        experience.ExperienceName = ExperienceName;
+        experience.ExperienceDescription = ExperienceDescription;
+        experience.ExperienceImage = ExperienceImage;
+        const updatedExp = await experience.save();
+        res.status(200).json(updatedExp);
     } catch (error) {
         console.error(error);
         res.status(500).json({ error: 'Unable to update experience' });
     }
 });
 
-// Delete an Experience (doesn't remove Experience from DayExperiences arrays)
-routerExperiences.delete('/:id', async (req, res) => {
+// Delete an Experience (removes Experience from DayExperiences arrays)
+routerExperiences.delete('/:id', checkJwt, async (req, res) => {
     try {
+        const experience = await Experience.findById(req.params.id);
+        const userId = req.auth.sub.split('|')[1];
+
+        if (!experience) {
+            return res.status(404).json({ 'Error': 'Experience not found' });
+        }
+        if (experience.userId !== userId) {
+            return res.status(403).json({ 'Error': 'Forbidden: You are not allowed to delete this experience' });
+        }
+
         await Experience.findByIdAndDelete(req.params.id);
         res.status(204).end();
     } catch (error) {
@@ -269,13 +299,17 @@ routerExperiences.delete('/:id', async (req, res) => {
 });
 
 // Add an Experience to a Day of a Trip
-routerTrips.post('/:tripId/:dayId/:expId', async (req, res) => {
+routerTrips.post('/:tripId/:dayId/:expId', checkJwt, async (req, res) => {
     try {
         const { tripId, dayId, expId } = req.params;
+        const userId = req.auth.sub.split('|')[1];
 
         const trip = await Trip.findById(tripId);
         if (!trip) {
             return res.status(404).json({ 'Error': 'Trip not found' });
+        }
+        if (trip.userId !== userId) {
+            return res.status(403).json({ 'Error': 'You are not authorized to add an experience to this trip' });
         }
 
         const day = trip.TripDays.find(day => day._id.toString() === dayId);
@@ -294,13 +328,17 @@ routerTrips.post('/:tripId/:dayId/:expId', async (req, res) => {
 });
 
 // Delete an Experience from a Day of a Trip (doesn't delete Experience)
-routerTrips.delete('/:tripId/:dayId/:expId', async (req, res) => {
+routerTrips.delete('/:tripId/:dayId/:expId', checkJwt, async (req, res) => {
     try {
         const { tripId, dayId, expId } = req.params;
+        const userId = req.auth.sub.split('|')[1];
 
         const trip = await Trip.findById(tripId);
         if (!trip) {
             return res.status(404).json({ 'Error': 'Trip not found' });
+        }
+        if (trip.userId !== userId) {
+            return res.status(403).json({ 'Error': 'You are not authorized to add an experience to this trip' });
         }
 
         const day = trip.TripDays.find(day => day._id.toString() === dayId);
@@ -324,35 +362,41 @@ routerTrips.delete('/:tripId/:dayId/:expId', async (req, res) => {
 });
 
 // Move an Experience from one Day of a Trip to another
-routerTrips.put('/:tripId/:dayId1/:dayId2/:expId', async (req, res) => {
+routerTrips.put('/:tripId/:dayId1/:dayId2/:expId', checkJwt, async (req, res) => {
     try {
-      const { tripId, dayId1, dayId2, expId } = req.params;
-  
-      const trip = await Trip.findById(tripId);
-      if (!trip) {
-        return res.status(404).json({ 'Error': 'Trip not found' });
-      }
-  
-      const day1 = trip.TripDays.find(day => day._id.toString() === dayId1);
-      const day2 = trip.TripDays.find(day => day._id.toString() === dayId2);
-      if (!day1 || !day2) {
-        return res.status(404).json({ 'Error': 'One or both days not found' });
-      }
-  
-      const index = day1.DayExperiences.indexOf(expId);
-      if (index === -1) {
-        return res.status(404).json({ 'Error': 'Experience not found in the specified day' });
-      }
-  
-      const experience = day1.DayExperiences.splice(index, 1)[0];
-      day2.DayExperiences.push(experience);
-  
-      await trip.save();
-  
-      res.status(200).json(trip);
+        const { tripId, dayId1, dayId2, expId } = req.params;
+        const userId = req.auth.sub.split('|')[1];
+        const trip = await Trip.findById(tripId);
+        if (!trip) {
+            return res.status(404).json({ 'Error': 'Trip not found' });
+        }
+        if (trip.userId !== userId) {
+            return res.status(403).json({ 'Error': 'Forbidden: You are not allowed to move experiences in this trip' });
+        }
+        if (dayId1 === dayId2) {
+            return res.status(400).json({ 'Error': 'Source and destination days cannot be the same'});
+        }
+        // Retrieve day info from TripDays array of trip and check their existence
+        const day1 = trip.TripDays.find(day => day._id.toString() === dayId1);
+        const day2 = trip.TripDays.find(day => day._id.toString() === dayId2);
+        if (!day1 || !day2) {
+            return res.status(404).json({ 'Error': 'One or both days not found' });
+        }
+
+        // Find index of experience in the DayExperiences array
+        const index = day1.DayExperiences.indexOf(expId);
+        if (index === -1) {
+            return res.status(404).json({ 'Error': 'Experience not found in the specified day' });
+        }
+        // Removes experience from day1 array 
+        const experience = day1.DayExperiences.splice(index, 1)[0];
+        // Adds experiences to day2 array
+        day2.DayExperiences.push(experience);
+        await trip.save();
+        res.status(200).json(trip);
     } catch (error) {
-      console.log(error);
-      res.status(500).json({ 'Error': 'Unable to move experience' });
+        console.log(error);
+        res.status(500).json({ 'Error': 'Unable to move experience' });
     }
   });
